requirements.txt
torch==2.0.1
torchaudio==2.0.2
hydra-core==1.3.2
lightning==2.0.4
wandb==0.15.4
transformers==4.30.2
jiwer==3.0.2
phonemizer==3.2.1